removed alerts to console, bugfixing
